Fix scheduled reminders not being cancelled on mark done or delete
Commit 666944223f10e32b7af21aa9354990ffdd91a12b added an action to the intent
used to schedule reminders with AlarmManager. However, the action was not added
to the intent used to cancel the corresponding alarm. As the intents were then
not considered equal anymore, the alarm was not cancelled when marking a
reminder done or deleting it.

Adding the same action to the intent used to cancel the alarm solved the problem.