Fix incorrect rescheduling of reminders leading to crash
Commit 666944223f10e32b7af21aa9354990ffdd91a12b introduced an
action added to intents scheduled with AlarmManager to prevent
loss of extras. Actually the loss of extras was probably due
to the reuse of old pending intents when rescheduling. Using
the action when scheduling but not when cancelling actually
caused the pending intents not to be cancelled correctly;
but as a side-effect, the intent used to cancel could not be
mistakenly reused either.

After cancelling was corrected in 7a3c121526d04b2f1f1a13123d4724fa49cc3210
by adding the action also to the intent used to cancel,
intents could be confused again and probably the intent
used to cancel was sometimes reused when rescheduling
reminders which led to crashes as ReminderService would
throw an exception when called without extras (which the
intent used to cancel does not have and need).

This is now fixed by using PendingIntent.FLAG_CANCEL_CURRENT
for each pending intent to start ReminderService.