New branch: develop
New development will happen on this branch.
Here the applicatioId has a suffix ".dev" so that the app will be installed
as a different app than the release version.
This is also reflected in the version string.
Version codes (and string) will now only increment when starting work on
a new version (and the same version code will be used on dev and release
channel as they are the corresponding versions for two separate apps).